-When importing, a vertex-only version of the mesh is created.
-This version is used when rendering shadows, and improves performance by reducing bandwidth
-It's automatic, but can optionally be used by users, in case they want to make special versions of geometry for shadow casting.

On thre 1060 this improves shadow rendering by 0.2 ms (from 2.4 to 2.2 ms in a test scene), on a R7 250 and Intel IGP this is a more significant improvement, from 7ms to 4.6ms.

Sorry, just an optimization, so nothing really visible.