@Calinou 2021-01-25
> Sorry, just an optimization, so nothing really visible.

No need to apologize. Sometimes, the best changes are the quietest ones :slightly_smiling_face: