@JFonS 2021-01-25
A bit tangential to this PR, but I feel like `EditorSceneImporterMesh::get_mesh()` has a misleading name. It doesn't convey the fact that it will create the mesh if it's not created already.

While testing, I was confused because `create_shadow_mesh()` was not being called for a GLTF model. It turns out the GLTF importer calls `get_mesh()` for an error check here: https://github.com/godotengine/godot/blob/c5c9517e1e03175e9cdc483d3dce27fc531af2be/modules/gltf/gltf_document.cpp#L5681

The check is not only useless, because get_mesh() never returns a null reference, but it causes an error.

The `get_mesh()` call generates the mesh early and prevents the `create_shadow_mesh()` call to take place  in `ResourceImporterScene::_generate_meshes()`, since it checks if the mesh is already generated.