@reduz 2021-01-25
@Zireael07 will probably work in GLES3 too