@reduz 2021-01-25
@JFonS API is not great, I think a proposal or a PR to improve it would be good. @JFonS that looks like a problem in the GLTF importer, @fire should give that a check.