@JFonS 2021-01-25
yup, It's definitely a bug in the gltf importer, but my point was that the API makes it very easy to make this kind of mistakes. Any call to `get_mesh()` before `ResourceImporterScene::_generate_meshses()` is called will cause issues. 

As I said, it was mostly unrelated to this PR, just pointing it out :)